Refactor: move updater.cpp from utils to source
